Move everything from Command to Service
